encra_pkg/setting/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Cython==3.0.12
gmpy2==2.2.1
pytest==8.3.5
pycryptodome==3.10.1
itsdangerous>=2.1.2
PyJWT==2.3.0
cryptography==3.4.7
Flask-Mail==0.9.1